Séparé le dropdown en deux components
